improvement, log the progress of creating the index.json file (#9434)
For large scopes, when index.json is deleted it takes a while to create
a new one (because it needs to read every object file). In the
meanwhile, there is no indication that something is happening. This PR
logs to the debug.log the progress every 1,000 files.
Also, the implementation of p-map-pool is simplified by using `chunk`
from `lodash`.