improvement, avoid deleting index.json when it has missing objects (#9435)
Currently, when an object (lane/component) exists in index.json but
missing from the filesystem, bit deletes the index.json assuming it's
corrupted and re-create it on the next command.
For large scopes, it takes long time to re-create the file because it
needs to parse each and every object file.
This PR approaches it differently. It updates the index.json file,
removes the missing objects and then throws an error telling the user to
re-run the command.