fix(tag-from-scope), avoid publishing artifacts in the package (#9468)
When tagging from scope, the artifacts from the last snap are written
into the capsule as part of the pipeline execution. The paths are the
same as they're saved in the model (`Version` object).
Currently, the preview/schema/test artifacts have relativePath without
the `artifacts/` prefix. It's happening due to the usage of the
deprecated `rootDir`.
This fix removes the `rootDir` from their artifact definition and use
`globPatterns` instead. The paths are now written correctly to the
model, and then when tagging from scope, they're written into
`artifacts` dir properly.

Another issue was the `.npmignore` which wasn't get written at all when
tagging from scope because it was part of the build pipeline not
snap/tag. This PR copies it to the snap/tag pipeline.