improvement(ipc-events), support publishing events directly for SSE (#9476)
This is helpful when all we want is to notify the watchers that a new
event has been written to the filesystem.
It doesn't require the watchers to register to GotEventSlot. The
watchers always listen to this onNotifySSE event, and once they get it,
they simply send a server-send-event to the clients.