fix(tag), fix buildStatus when tagging with --build aspects (#9477)
In some cases, when tagging or snapping with `--build` flag, the
buildStatus remain `pending`, instead of `succeed`.
This is a regression caused by https://github.com/teambit/bit/pull/9437.
It happens for example when tagging/snapping aspects.
It's happening because the `ConsumerComponent instance inside the
harmony component is different than the one originally created. Although
the harmony component is created by the legacy, but in case the
`Workspace` aspect already had this component in the cache, it won't use
the provided legacy, but create a new one.
As a result, the buildStatus was set to one instance of
ConsumerComponent and another instance was saved into the filesystem.

This PR fixes it by making sure it's using the same instance of
`ConsumerComponent`.