improvement(snap), enable auto-formatting components before snap/tag (#9497)
Added a new slot `onPreSnap` for snapping aspect, which the formatter
aspect registered to. Before snapping the components, the formatter
checks the config of `formatOnPreSnap` and if enabled it formats the
modified components. It skips the auto-tag components as they don't have
code modifications.

The reason for not adding it as a build-task is because it makes changes
to the source files. It doesn't add metadata as any other build task. We
want it to run locally according to the user workspace.jsonc config and
regardless whether the user is running the build on the cloud or not.

To enable it, simply add the following to your workspace.jsonc file:
```
"teambit.defender/formatter": {
    "formatOnPreSnap": true
  }
```