fix(merge-from-scope), save dependencies data after merging config (#9511)
In case of merging main into an out-of-date lane, we compare the config
and if there is no conflict, we generate `.bit/unmerged.json` file with
the config to update.
When running from the workspace, as soon as the components are re-loaded
to be snapped during the process, it picks up this unmerged file and
combine the config from this file into the aspect data of the componet.
However, when running from the scope, there is no concept of
aspect-merge from different sources. As a result, the config from this
file are not integrated into the component.
This PR fixes it by loading the unmerged file and adding the data to the
component object.