replace env of core teambit.harmony/node with teambit.node/envs/node-babel-mocha (#9503)
whenever possible and upgrade the babel-mocha with to the latest
version.
This version has a babel config targets node 22 and many old plugins
were removed, as a result, the dists files are much more readable and
clean.
The directory "excluded-fixtures" has been set to be excluded from
compilation, types and tests.
Also, the "preview" task has been removed from this env. It's not needed
for these node components and they cause an issue when Webpack tries to
compile the "excluded-fixtures" directory.