change teambit.toolbox/* env to a new one: node-typescript-mocha (#9529)
Currently, they use node-babel-mocha, which includes the lazy loading
plugin. The problem is that this plugin wraps all require/import
statements in a function, and when a named export is called from an ESM
file, Node throws an error. In the new environment, compilation is
performed using TypeScript instead of Babel, so there are no issues when
using them from an ESM file.