fix(checkout-reset), write deleted main-component when on a lane (#9533)
`bit checkout reset --all` rewrites deleted component dirs.
Currently, when running it and a non-lane component is deleted when the
workspace is on a lane, this component is not written.
The reason is that when on a lane, the ids used by this command are the
ids from the lane. While this is the expected behavior for `bit checkout
head`, it's not the case for `bit checkout reset`.
This PR changes it so that all workspace ids are used for the
checkout-reset.