fix(new), fix env-not-found regression error by not requiring the env to be loaded (#9544)
This is a regression caused by https://github.com/teambit/bit/pull/9530.
That API needed the env to be loaded when writing the package.json. In
some cases, this is hard to achieve. For `bit new`, when the env is not
there, it threw an error:

> environment with ID: X configured on component Y was not found

This PR fixes it by not throwing an error in such case.