improvement(merge-from-scope), skip loading aspects when merging main to a lane (#9570)
When merging main into a lane from scope, it has to make a new snap with
2 parents: main and lane. During the snap, we used to load aspects
because when trying to avoid the aspect-loading, aspects data weren't
saved correctly.
The reason is that aspects-data gets saved on the component object
during load. If aspect-loading doesn't happen, the env is not loaded and
then when the component is loaded, it doesn't find the env and fallback
to the default node-env.

This PR introduces a new mechanism of merging aspects-data. In case both
sides: lane and main has the same env, we can simply compare the
aspect-data and if there is no conflict, we can keep or copy the data as
needed.
Eventually, before snap, we check which components were merged their
data properly and for them, we don't load aspects. In case of data
conflicts or env changes we do load aspects.