perf, lazy-load component-templates (#9577)
Currently, many templates are loaded during Bit’s bootstrap process.
This occurs because aspects call `generator.registerComponentTemplate`
inside their `provider` function, which is invoked whenever the aspect
is loaded.
It's addressed by changing the `registerComponentTemplate` to accept
also a function that returns the templates. This way, the templates are
retrieved only when needed.
This change reduces the number of files loaded during bootstrap by 229.