improvement(create), provide an infra for interactive templates (#9610)
## Proposed Changes

- Provide an API for component-templates creator `promptOptions`. If
implemented, `bit create` will show the prompts, gather the info needed,
and then pass this info to the `generateFiles()` of the template.
- Support adding the prompt-options as flags. (e.g. bit create comp
some-template -- --bundler webpack).
- This infrastructure eliminates the need from templates creator to
import a prompt library and implement all of it themselves.
- The main reason of implementing this is to enable the IDE extension to
display prompts within the IDE UI and create the components according to
the user's input.