fix(merge-from-scope), fix VersionNotFound error when merging main into a lane in some edge cases (#9629)
In some cases, when merging main to a lane from a bare-scope, it throws
an error about not finding the Version object of a squashed one in the
history.
This is happening when exporting the lane to the remote. Bit tries to
export the missing history from main into the lane's scope.
Due to recent changes in the remote, it's not needed anymore, and the
remote itself fetches the missing history once the lane is exported.
Instead, this PR, only pushes the head into the lane's scope, and
doesn't try to fetch missing history.