fix(merge-from-scope), enable auto-snap when merging main into a lane (#9646)
When merging main into a lane, the merged components are snapped. This
PR finds the other components in the lane that depend on the snapped
components and auto-snap them. This way, the lane is kept up-to-date,
and components in the lane always have the dependencies with the same
versions they appear on the lane.