fix(dep-resolver), validate deps-resolver aspect data (specifically "version" field) before saving it (#9648)
Currently, the "env.jsonc" can have any string in the "version" field,
and it'll be saved into the objects incorrectly.
This PR makes the following validations:
1. "version" field needs to be a semver/semver-range valid.
2. for components, it can be a hash, in case the version is a snap.
3. the `*` range is valid only for packages, not for components. The
reason is that for components, in case it was snapped on a lane, the
package on bit cloud doesn't have "latest" tag, as a result, the
installer fails.