deprecate "test" and "name" properties of the files in the model (#9665)
The `test` prop is not used since Harmony. The `name` is easily
calculated by `path.basename` of the `relativePath`, and it not actually
used.
The reason not to simply delete them is to be forward-compatible.
Meaning, a component that was tagged with the new version won't appear
as "modified" by an older version.
The "modified" calculation has been changed to disregard these two
props.