avoid printing "Events were dropped by the FSEvents client" during watch (#9678)
For big projects, it happens all the time when running `bit install` or
any heavy command. It's because Parcel watches the workspace root and
although the node-modules is ignored, it probably stills gets the events
from the FSEvents (and then discard them).
It's not helpful to show this error message constantly, so it's simply
ignored.