fix: avoid error ComponentsPendingImport when bit-watch or IDE extension starts (#9863)
Fixes the issue where `bit watch` or the IDE extension shows "your
workspace has outdated objects. please use 'bit import' to pull the
latest objects from the remote scope".
This is happening when the user is running `git pull`, which updates
`.bitmap` file, but the local scope objects are not updated yet ("bit
import" is not run yet).
To fix it, this PR checks before those two operations whether there are
out of date components against .bitmap, and if it found some, it fetches
the objects for them.