fix(ci-merge): reset components that were created on a lane to new after switching to main (#9896)
Here is the use-case where this bug occurred - a user is running `bit
lane import --branch`, which creates a git-branch with the lane data,
`bit ci pr` snaps and exports the changes on the lane. Once done, the
`bit ci merge` switches to main and tries to tag. However, because all
changes were done on the lane, once switched to main, those components
are marked as unavailable.
This PR fixes it by running "bit init --reset-lane-new` programmatically
after switching to main. This way, the components that were created on
the lane become new components.